Fix comparison warnings in fractal kernels. Modify CI scripts: brewfile, add x11 package to linux addons.

+++ examples/julia/include/kernels/julia.vanilla.hpp
@@ -64,7 +64,7 @@ namespace zacc { namespace examples {
         {
             vec2<float> center = _dim / 2.0;
 
-            for(int i = 0; i < output.size(); i++)
+            for(size_t i = 0; i < output.size(); i++)
             // populate output container
             //zacc::generate<int>(std::begin(output), std::end(output), [this](int i)
             {
